requirements/common.txt
bson==0.5.8
graphene==2.1.6
python-dateutil==2.8.0
python-multipart==0.0.5
pytz==2019.1
PyYAML==5.1.1
readme-renderer==24.0
requests-toolbelt==0.9.1
requests==2.22.0
singledispatch==3.4.0.3
six==1.12.0
toml==0.10.0
tqdm==4.34.0
twine==1.13.0
typing==3.7.4
